Create task-43.md
Translating this one was really difficult. One thing is to understand the task and the second is to translate it into human language
and not to look stupid while doing so
which is in cases like "floating point" really difficult. I bet that russian language has similar issue so I think you could relate :grin: