About dozen tasks in SK
* Create task-30.md

* Create task-17.md

* Delete task-17.md

* Create task-17.md

* Create task-27.md

* Create task-156.md

* Create task-215.md

* Update task-215.md

Seems like there is no SK or CZ wiki explaining klotski puzzle :(

* Update task-215.md

* Create task-94.md

Not much of a work translating this one was. (:

* Update task-1.md

* Create task-50.md

* Create task-26.md

I added link to sk wiki https://sk.wikipedia.org/wiki/Euklidov_algoritmus I hope it's not a problem

* Create task-47.md

* Update task-27.md

* Create task-43.md

Translating this one was really difficult. One thing is to understand the task and the second is to translate it into human language
and not to look stupid while doing so
which is in cases like "floating point" really difficult. I bet that russian language has similar issue so I think you could relate :grin:

* Update task-43.md

* Create task-10.md

* Create task-21.md

* Update task-20.md

Minor mistake fixed.

* Update task-11.md